feat: separate files and add abc's
